SignUpコンポーネント大雑把に作成
